Make the default board non-deletable & go back to main board after board deletion

+++ client/src/main/java/client/scenes/MainSceneCtrl.java
@@ -9,12 +9,14 @@ import commons.TaskList;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
+import javafx.stage.Modality;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -134,9 +136,16 @@ public class MainSceneCtrl {
             System.out.println("Cannot delete board: this is a dummy board!");
             return;
         }
-        mainCtrl.setActiveBoard(null);
+        if (board.getId() == 1030) {
+            var alert = new Alert(Alert.AlertType.ERROR);
+            alert.initModality(Modality.APPLICATION_MODAL);
+            alert.setContentText("You cannot delete the default board!");
+            alert.showAndWait();
+            return;
+        }
+        mainCtrl.setActiveBoard(server.getDefaultBoard());
         server.deleteBoard(board);
-        mainCtrl.showConnect();
+        refresh();
     }
 
     /**
